requirements.txt
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
reportlab>=4.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
pillow>=10.2.0
pydantic>=2.5.0
